requirements.txt

beautifulsoup4~=4.9.3
cryptography==2.8
Cython~=0.29.14





Django==3.0.5
djangorestframework==3.12.1

djongo==1.3.3
gensim~=3.8.0
google==3.0.0
nltk~=3.5
numpy~=1.19.2
pymongo==3.11.0
requests==2.24.0
tweepy==3.9.0








pip~=20.0.2
zlib~=1.2.11
wheel~=0.34.2
openssl~=1.1.1h
py~=1.8.1
lxml~=4.5.0
keyring~=21.1.0
setuptools~=45.2.0
textblob~=0.15.3



pytest~=6.1.1
spacy~=2.3.2
typing~=3.7.4.3

jsonfield~=3.1.0
readability-lxml~=0.8.1
trafilatura~=0.6.0

pytrends~=4.7.3
dowhy~=0.5
yfinance~=0.1.5
xlrd~=1.2.0